+++ src/main/java/com/contentful/java/cma/model/CMAType.java
@@ -11,6 +11,7 @@ public enum CMAType {
   Entry,
   Link,
   Locale,
+  Organization,
   Role,
   Space,
   SpaceMembership,